refactor: move book controller logic into router file
